+++ .goreleaser.windows.yml
@@ -16,6 +16,8 @@ builds:
     goarch:
       - "amd64"
     mod_timestamp: "{{ .CommitTimestamp }}"
+    tags:
+      - "memoryprotection"
     ldflags:
       - "-s -w -checklinkname=0"
       - "-X github.com/jzelinskie/cobrautil/v2.Version=v{{ .Version }}"
