+++ cmd/spicedb/integration/main_test.go
@@ -1,4 +1,4 @@
-package main
+package integration_test
 
 import (
 	"bytes"
