+++ cmd/spicedb/integration/servetesting_integration_test.go
@@ -1,6 +1,6 @@
 //go:build docker && image
 
-package main
+package integration_test
 
 import (
 	"bytes"
