+++ cmd/spicedb/integration/testdata/bootstrap.yaml
