+++ cmd/spicedb/main.go
@@ -9,6 +9,7 @@ import (
 	"google.golang.org/grpc/balancer"
 	_ "google.golang.org/grpc/xds"
 
+	"github.com/authzed/spicedb/cmd/spicedb/memoryprotection"
 	log "github.com/authzed/spicedb/internal/logging"
 	"github.com/authzed/spicedb/pkg/cmd"
 	cmdutil "github.com/authzed/spicedb/pkg/cmd/server"
@@ -17,6 +18,8 @@ import (
 )
 
 func main() {
+	memoryprotection.InitDefaultMemoryUsageProvider()
+
 	// Set up root logger
 	// This will typically be overwritten by the logging setup for a given command.
 	log.SetGlobalLogger(zerolog.New(os.Stderr).Level(zerolog.InfoLevel))
