+++ magefiles/gen_docs.go
@@ -9,7 +9,7 @@ import (
 
 // Docs Generate documentation in markdown format
 func (g Gen) Docs() error {
-	cmd := exec.Command("go", "run", "-ldflags=-s -w -checklinkname=0", "./cmd/docs")
+	cmd := exec.Command("go", "run", "./cmd/docs")
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 	return cmd.Run()
